+++ backend/src/models/users.ts
@@ -1,48 +1,71 @@
 import db from '../schemas';
 import { User, SequelizeUser } from '../schemas/users';
+import { City } from '../schemas/cities';
 
 /**
  * Get all items on the table without any filter
+ * @param cityId logged user city ID
  * @returns Promise<List of items>
  */
-export const getAll = (): Promise<SequelizeUser[]> => {
-  return db.users.findAll();
+export const getAll = (cityId: NonNullable<City['id']>): Promise<SequelizeUser[]> => {
+  return db.users.findAll({ where: { cityId } });
 };
 
 /**
  * Get a single item using the unique ID
  * @param id unique ID of the desired item
+ * @param cityId logged user city ID
  * @returns Promise<Item>
  */
-export const getById = (id: string | number): Promise<SequelizeUser> => {
-  return db.users.findByPk(id);
+export const getById = async (
+  id: NonNullable<User['id']>,
+  cityId: NonNullable<City['id']>
+): Promise<SequelizeUser | null> => {
+  const item = await db.users.findByPk(id);
+  if (item?.cityId === cityId) return item;
+  return null;
 };
 
 /**
  * Function to create a new row on the table
  * @param values object with the new item data
+ * @param cityId logged user city ID
  * @returns Promise<Item>
  */
-export const create = (values: User | SequelizeUser): Promise<SequelizeUser> => {
-  return db.users.create(values);
+export const create = (values: User | SequelizeUser, cityId: NonNullable<City['id']>): Promise<SequelizeUser> => {
+  return db.users.create({ ...values, cityId });
 };
 
 /**
  * Function to update a row on the table by the unique ID
  * @param id unique ID of the desired item
  * @param values object with the new data
+ * @param cityId logged user city ID
  * @returns Promise<Item>
  */
-export const updateById = async (id: string | number, values: User | SequelizeUser): Promise<SequelizeUser> => {
-  // The update return an array [count, item[]], so I'm destructuring to get the updated user
-  const [, [item]] = await db.users.update(values, { where: { id }, returning: true });
-  return item;
+export const updateById = async (
+  id: NonNullable<User['id']>,
+  values: User | SequelizeUser,
+  cityId: NonNullable<City['id']>
+): Promise<SequelizeUser | null> => {
+  // Check if item is on the city
+  const cityItem = getById(id, cityId);
+  if (cityItem) {
+    // The update return an array [count, item[]], so I'm destructuring to get the updated user
+    const [, [item]] = await db.users.update(values, { where: { id }, returning: true });
+    return item;
+  }
+  return null;
 };
 
 /**
  * Function to delete a row on the table by the unique ID
- * @param id unique ID of the desired item
+ * @param id unique ID of the desired item *
+ * @param cityId logged user city ID
  */
-export const deleteById = (id: string | number): void => {
-  db.users.destroy({ where: { id } });
+export const deleteById = async (id: NonNullable<User['id']>, cityId: NonNullable<City['id']>): Promise<void> => {
+  const cityItem = getById(id, cityId);
+  if (cityItem) {
+    await db.users.destroy({ where: { id } });
+  }
 };
